Update fi.json
Hello! I am a Finnish King of the Hill - Cosmicv player who has noticed that there are shortcomings in the translations related to the Finnish language on the server. Some words are not translated into Finnish at all, and in some translations, the Finnish is poor. Here is my own translation! I hope it will be useful :)