Update fi.json (#50)
* Update fi.json

Hello! I am a Finnish King of the Hill - Cosmicv player who has noticed that there are shortcomings in the translations related to the Finnish language on the server. Some words are not translated into Finnish at all, and in some translations, the Finnish is poor. Here is my own translation! I hope it will be useful :)

* Update fi.json

"I corrected a few typos.

* Update fi.json

Line 178-180 fixed

* fix json formatting

---------

Co-authored-by: Paul Reed <45607581+TheCoolestPaul@users.noreply.github.com>